tnajam: Move fragment classes to the corresponding directory
