+++ app/src/main/java/com/udacity/shoestore/fragments/ShoeDetailFragment.kt
@@ -1,10 +1,11 @@
-package com.udacity.shoestore
+package com.udacity.shoestore.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.udacity.shoestore.R
 
 class ShoeDetailFragment : Fragment() {
