+++ app/src/main/res/layout/fragment_instructions.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".InstructionsFragment">
+    tools:context=".fragments.InstructionsFragment">
 
     <TextView
         android:id="@+id/tv_instructions_heading"
