+++ app/src/main/res/layout/fragment_shoe_detail.xml
@@ -27,7 +27,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:orientation="vertical"
-            tools:context=".ShoeDetailFragment">
+            tools:context=".fragments.ShoeDetailFragment">
 
 
             <ImageView
